bot_candidates/formula-trend/02/requirements.txt
h5py==2.8.0
Keras==2.1.6
Pillow==4.0.0
tensorflow==1.10.0
flask-socketio
eventlet